Feat: ADD 문장 감성점수 계산, 문장 추천 기능 추가
Former-commit-id: f83f04e5ed7fd9153eb40050473caee5fcb60415

+++ BE/django/recommend-service/basic_recom/models.py
@@ -112,7 +112,7 @@ class Paragraph(models.Model):
     paragraph_page = models.IntegerField()
     scrap_count = models.IntegerField(blank=True, null=True)
     user = models.ForeignKey('Userr', models.DO_NOTHING)
-    paragraph_score = models.IntegerField(blank=True, null=True)
+    paragraph_score = models.FloatField(blank=True, null=True)
     paragraph_score_type = models.CharField(max_length=45, blank=True, null=True)
 
     class Meta:
